+++ Assets/Scripts/Vegetables/TomatoClass.cs
@@ -9,7 +9,7 @@ public class TomatoClass : VegetableClass
 
     }
 
-    public TomatoClass(int health, int damage, float speed, string prefab, GameObject gameObject) 
+    public TomatoClass(int health, float damage, float speed, string prefab, GameObject gameObject) 
         : base(health, damage, speed, prefab, gameObject)
     {
